+++ Runtime/Helpers/StatusEventSender.cs
@@ -1,11 +1,14 @@
 using System;
 using UnityEngine;
 
-public class StatusEventSender : MonoBehaviour
+namespace Mane
 {
-    public event Action<GameObject, bool> StatusChanged;
+    public class StatusEventSender : MonoBehaviour
+    {
+        public event Action<GameObject, bool> StatusChanged;
 
-    private void OnEnable() => StatusChanged?.Invoke(gameObject, true);
+        private void OnEnable() => StatusChanged?.Invoke(gameObject, true);
 
-    private void OnDisable() => StatusChanged?.Invoke(gameObject, false);
+        private void OnDisable() => StatusChanged?.Invoke(gameObject, false);
+    }
 }