+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.10.10:
+    - Added GetRequiredComponent for the Component class, to call directly from the component w/o using gameObject.
+    - Added Mane.UI namespace, a bit reorganised UI components.
+
 1.10.9:
     - Improved Color Scheme Component inspector view.